+++ src/core/meson.build
@@ -32,6 +32,7 @@ endif
 nixl_lib = library('nixl',
                    'nixl_agent.cpp',
                    'nixl_plugin_manager.cpp',
+                   'nixl_listener.cpp',
                    include_directories: [ nixl_inc_dirs, utils_inc_dirs ],
                    dependencies: nixl_lib_deps,
                    install: true)
